refactor: merge providers (#241)
