+++ src/provider/providerWrapper.ts
@@ -39,11 +39,15 @@ interface GetDataReturn {
   value: Record<string, any> | null;
 }
 
-export class Web3ProviderWrapper {
+export class ProviderWrapper {
   type: ProviderTypes;
   provider: any;
   constructor(provider: any) {
-    this.type = ProviderTypes.WEB3;
+    if (typeof provider.request === 'function') {
+      this.type = ProviderTypes.ETHEREUM;
+    } else {
+      this.type = ProviderTypes.WEB3;
+    }
     this.provider = provider;
   }
 
@@ -106,7 +110,9 @@ export class Web3ProviderWrapper {
         `${interfaceId}${'00000000000000000000000000000000000000000000000000000000'}`,
       ),
     );
-
+    if (this.type === ProviderTypes.ETHEREUM) {
+      return decodeResult(Method.SUPPORTS_INTERFACE, result);
+    }
     return decodeResult(Method.SUPPORTS_INTERFACE, result.result);
   }
 
@@ -118,6 +124,23 @@ export class Web3ProviderWrapper {
    * @param signature
    */
   async isValidSignature(address: string, hash: string, signature: string) {
+    if (this.type === ProviderTypes.ETHEREUM) {
+      const encodedParams = abiCoder.encodeParameters(
+        ['bytes32', 'bytes'],
+        [hash, signature],
+      );
+
+      const result = await this.callContract(
+        constructJSONRPC(address, Method.IS_VALID_SIGNATURE, encodedParams),
+      );
+
+      if (result.error) {
+        throw result.error;
+      }
+
+      return decodeResult(Method.IS_VALID_SIGNATURE, result);
+    }
+
     const encodedParams = abiCoder.encodeParameters(
       ['bytes32', 'bytes'],
       [hash, signature],
@@ -135,7 +158,6 @@ export class Web3ProviderWrapper {
 
   async getData(address: string, keyHash: string) {
     const result = await this.getAllData(address, [keyHash]);
-
     try {
       return result[0].value;
     } catch {
@@ -169,6 +191,23 @@ export class Web3ProviderWrapper {
     address: string,
     keyHashes: string[],
   ): Promise<GetDataReturn[]> {
+    if (this.type === ProviderTypes.ETHEREUM) {
+      const encodedResults = await this.callContract(
+        constructJSONRPC(
+          address,
+          Method.GET_DATA,
+          abiCoder.encodeParameter('bytes32[]', keyHashes),
+        ),
+      );
+
+      const decodedValues = decodeResult(Method.GET_DATA, encodedResults);
+
+      return keyHashes.map<GetDataReturn>((keyHash, index) => ({
+        key: keyHash,
+        value: decodedValues[index],
+      }));
+    }
+
     const payload: JsonRpc[] = [
       constructJSONRPC(
         address,
@@ -190,8 +229,24 @@ export class Web3ProviderWrapper {
     address: string,
     keyHashes: string[],
   ): Promise<GetDataReturn[]> {
-    const payload: JsonRpc[] = [];
+    if (this.type === ProviderTypes.ETHEREUM) {
+      // Here we could use `getDataMultiple` instead of sending multiple calls to `getData`
+      // But this is already legacy and it won't be used anymore..
+      const encodedResultsPromises = keyHashes.map((keyHash) =>
+        this.callContract(
+          constructJSONRPC(address, Method.GET_DATA_LEGACY, keyHash),
+        ),
+      );
 
+      const decodedResults = await Promise.all(encodedResultsPromises);
+
+      return decodedResults.map((decodedResult, index) => ({
+        key: keyHashes[index],
+        value: decodeResult(Method.GET_DATA_LEGACY, decodedResult),
+      }));
+    }
+
+    const payload: JsonRpc[] = [];
     // Here we could use `getDataMultiple` instead of sending multiple calls to `getData`
     // But this is already legacy and it won't be used anymore..
     for (let index = 0; index < keyHashes.length; index++) {
@@ -212,6 +267,13 @@ export class Web3ProviderWrapper {
   }
 
   private async callContract(payload: JsonRpc[] | JsonRpc): Promise<any> {
+    if (this.type === ProviderTypes.ETHEREUM) {
+      return this.provider.request({
+        method: 'eth_call',
+        params: (payload as JsonRpc).params,
+      });
+    }
+
     return new Promise((resolve, reject) => {
       // Send old web3 method with callback to resolve promise
       // This is deprecated: https://docs.metamask.io/guide/ethereum-provider.html#ethereum-send-deprecated
